backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary>=2.9.9

# LLM
ollama>=0.3.0
httpx>=0.25.0

# SSH
paramiko==3.4.0
cryptography==42.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring
psutil==5.9.8

# WebSockets
websockets==12.0

# Logging
structlog==24.1.0

# Retry logic
tenacity==8.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9
